fix(arc-154): fixes for navigation module, fix circular deps. (wip)
