+++ demo/src/modules/user/user.tsx
@@ -1,10 +1,10 @@
-import { FC, ReactNode } from "react";
-import { Route } from "react-router-dom";
+import { FC, ReactNode } from 'react';
+import { Route } from 'react-router-dom';
 
-import { USER_PATH } from "~modules/shared/consts/user.const";
-import UserOverview from "~modules/user/views/UserOverview";
-import { AdminLayout } from "~modules/shared/layouts";
-import { useTranslation } from "react-i18next";
+import { USER_PATH } from '~modules/shared/consts/user.const';
+import { UserOverview } from '~modules/user/views/UserOverview';
+import { AdminLayout } from '~modules/shared/layouts';
+import { useTranslation } from 'react-i18next';
 
 const UserOverviewPage: FC = () => {
 	const { t } = useTranslation();
@@ -15,8 +15,8 @@ const UserOverviewPage: FC = () => {
 				<UserOverview />
 			</AdminLayout.Content>
 		</AdminLayout>
-	)
-}
+	);
+};
 
 export const renderAdminUserRoutes = (): ReactNode[] => {
 	return [
@@ -28,3 +28,4 @@ export const renderAdminUserRoutes = (): ReactNode[] => {
 		/>,
 	];
 };
+
