+++ demo/src/react-admin/modules/content-page/components/ContentBlockPreview/ContentBlockPreview.tsx
@@ -3,7 +3,7 @@ import clsx from 'clsx';
 import { get, noop, omit } from 'lodash-es';
 import React, { FunctionComponent, RefObject, useCallback, useEffect, useRef } from 'react';
 
-import { GET_DARK_BACKGROUND_COLOR_OPTIONS } from '../../const/content-block.consts';
+import { GET_DARK_BACKGROUND_COLOR_OPTIONS } from '../../const/content-block.common.consts';
 import { Color, ContentBlockConfig } from '../../types/content-block.types';
 import { ContentPageInfo } from '../../types/content-pages.types';
