+++ demo/src/react-admin/modules/content-page/components/blocks/eventbrite.ts
@@ -1,7 +1,7 @@
 import { BlockEventbriteProps } from '@viaa/avo2-components';
 
 import { GET_ADMIN_ICON_OPTIONS } from '../../../shared/consts/icons.consts';
-import { GET_BUTTON_TYPE_OPTIONS } from '../../const/content-block.consts';
+import { GET_BUTTON_TYPE_OPTIONS } from '../../const/content-block.common.consts';
 import {
 	ContentBlockConfig,
 	ContentBlockEditor,
