+++ demo/src/react-admin/modules/content-page/components/blocks/image-grid.ts
@@ -2,11 +2,10 @@ import { GridItem } from '@viaa/avo2-components';
 
 import { FileUploadProps } from '../../../shared/components/FileUpload/FileUpload';
 import {
-	GET_ALIGN_OPTIONS,
 	GET_BUTTON_TYPE_OPTIONS,
 	GET_FILL_OPTIONS,
 	GET_IMAGE_GRID_FORMAT_OPTIONS,
-} from '../../const/content-block.consts';
+} from '../../const/content-block.common.consts';
 import {
 	ContentBlockConfig,
 	ContentBlockEditor,
@@ -19,6 +18,7 @@ import { BlockImageGridWrapperProps } from '../wrappers/ImageGridWrapper/ImageGr
 import { BLOCK_FIELD_DEFAULTS, BLOCK_STATE_DEFAULTS, FILE_FIELD, TEXT_FIELD } from './defaults';
 
 import { Config } from '~core/config';
+import { GET_ALIGN_OPTIONS } from '~modules/content-page/const/content-block.common.consts';
 
 export const INITIAL_IMAGE_GRID_COMPONENTS_STATE = (): ImageGridBlockComponentStateFields[] => [
 	{
