+++ demo/src/react-admin/modules/content-page/components/blocks/media-player-title-text-button.ts
@@ -7,7 +7,7 @@ import { validateFlowplayerVideoUrl } from '../../../shared/helpers/validation';
 import {
 	GET_BUTTON_TYPE_OPTIONS,
 	GET_HEADING_TYPE_OPTIONS,
-} from '../../const/content-block.consts';
+} from '../../const/content-block.common.consts';
 import {
 	ContentBlockConfig,
 	ContentBlockEditor,
