+++ demo/src/react-admin/modules/content-page/components/blocks/page-overview.ts
@@ -4,7 +4,7 @@ import {
 	GET_PAGE_OVERVIEW_ITEM_STYLE_OPTIONS,
 	GET_PAGE_OVERVIEW_ORDER_OPTIONS,
 	GET_PAGE_OVERVIEW_TAB_STYLE_OPTIONS,
-} from '../../const/content-block.consts';
+} from '../../const/content-block.common.consts';
 import {
 	Color,
 	ContentBlockConfig,
