+++ demo/src/react-admin/modules/content-page/helpers/get-published-state.ts
@@ -1,13 +1,9 @@
 import { Avo } from '@viaa/avo2-types';
-import { compact, get } from 'lodash-es';
+import { get } from 'lodash-es';
 import moment from 'moment';
 
-import { CONTENT_BLOCK_CONFIG_MAP } from '../const/content-block.consts';
-import { ContentBlockConfig, ContentBlockType } from '../types/content-block.types';
+import { ContentBlockConfig } from '../types/content-block.types';
 import { ContentPageInfo, PublishOption } from '../types/content-pages.types';
-
-import { Config, ToastType } from '~core/config';
-import { CustomError } from '~modules/shared/helpers/custom-error';
 
 export function getPublishedState(
 	contentPage: ContentPageInfo | Partial<ContentPageInfo> | undefined | null
@@ -96,57 +92,3 @@ export const convertBlocksToDatabaseFormat = (
 		convertBlockToDatabaseFormat(contentBlockConfig, contentId)
 	);
 
-// Parse content-blocks to configs
-export const parseContentBlocks = (
-	contentBlocks: Avo.ContentPage.Block[]
-): ContentBlockConfig[] => {
-	const sortedContentBlocks = contentBlocks.sort((a, b) => a.position - b.position);
-
-	return compact(
-		(sortedContentBlocks || []).map((contentBlock) => {
-			const { content_block_type, id, variables } = contentBlock;
-			const configForType = CONTENT_BLOCK_CONFIG_MAP[content_block_type as ContentBlockType];
-			if (!configForType) {
-				console.error(
-					new CustomError('Failed to find content block config for type', null, {
-						content_block_type,
-						contentBlock,
-						CONTENT_BLOCK_CONFIG_MAP,
-					})
-				);
-				Config.getConfig().services.toastService.showToast({
-					title: Config.getConfig().services.i18n.t(
-						'modules/admin/content-page/helpers/get-published-state___error'
-					),
-					description: Config.getConfig().services.i18n.t(
-						'modules/admin/content-page/helpers/get-published-state___er-ging-iets-mis-bij-het-laden-van-de-pagina'
-					),
-					type: ToastType.ERROR,
-				});
-				return null;
-			}
-			const cleanConfig = configForType(contentBlock.position);
-
-			const rawComponentState = get(variables, 'componentState', null);
-			const componentState = Array.isArray(rawComponentState)
-				? rawComponentState
-				: { ...cleanConfig.components.state, ...rawComponentState };
-
-			return {
-				...cleanConfig,
-				id,
-				components: {
-					...cleanConfig.components,
-					state: componentState,
-				},
-				block: {
-					...cleanConfig.block,
-					state: {
-						...cleanConfig.block.state,
-						...get(variables, 'blockState', {}),
-					},
-				},
-			} as ContentBlockConfig;
-		})
-	);
-};
