+++ demo/src/react-admin/modules/index.ts
@@ -1,4 +1,4 @@
 export * from './content-page';
 export * from './translations';
-export * from './user';
+export { UserOverview } from './user/views/UserOverview';
 export * from './user-group';
