+++ demo/src/react-admin/modules/navigation/queries/hetarchief/delete-navigation-item.graphql
@@ -1,0 +1,5 @@
+mutation deleteNavigationItem($id: uuid!) {
+	delete_app_navigation(where: { id: { _eq: $id } }) {
+		affected_rows
+	}
+}
