+++ demo/src/react-admin/modules/navigation/queries/hetarchief/get-navigation-elements.graphql
@@ -1,0 +1,8 @@
+query getNavigationElements {
+	app_navigation(distinct_on: placement, order_by: { placement: asc }) {
+		id
+		description
+		placement
+		tooltip
+	}
+}
