+++ demo/src/react-admin/modules/navigation/queries/hetarchief/get-navigation-item-by-id.graphql
@@ -1,0 +1,17 @@
+query getNavigationItemById($id: uuid!) {
+	app_navigation(where: { id: { _eq: $id } }) {
+		id
+		created_at
+		description
+		user_group_ids
+		icon_name
+		label
+		link_target
+		placement
+		position
+		updated_at
+		content_type
+		content_path
+		tooltip
+	}
+}
