+++ demo/src/react-admin/modules/navigation/queries/hetarchief/get-navigation-items-by-placement.graphql
@@ -1,0 +1,20 @@
+query getNavigationItemsByPlacement($placement: String!) {
+	app_navigation(
+		order_by: { position: asc }
+		where: { placement: { _eq: $placement } }
+	) {
+		id
+		created_at
+		description
+		user_group_ids
+		icon_name
+		label
+		link_target
+		placement
+		position
+		updated_at
+		content_type
+		content_path
+		tooltip
+	}
+}
