+++ demo/src/react-admin/modules/navigation/queries/hetarchief/update-navigation-item-by-id.graphql
@@ -1,0 +1,5 @@
+mutation updateNavigationItemById($id: uuid!, $navigationItem: app_navigation_set_input!) {
+	update_app_navigation(where: { id: { _eq: $id } }, _set: $navigationItem) {
+		affected_rows
+	}
+}
