+++ demo/src/react-admin/modules/navigation/views/NavigationDetail.scss
@@ -1,6 +1,5 @@
-@import '../../../styles/settings/colors';
-@import '../../../styles/settings/variables';
-
+@import "../../shared/styles/settings/colors";
+@import "../../shared/styles/settings/variables";
 
 // Elements
 .c-menu-detail__table {
