+++ demo/src/react-admin/modules/shared/consts/admin.const.ts
@@ -1,7 +1,7 @@
-import { NAVIGATION_PATH } from "~modules/navigation/navigation.const";
-import { COLLECTIONS_OR_BUNDLES_PATH } from "./collection.const";
-import { CONTENT_PATH } from "./content.const";
-import { USER_PATH } from "./user.const";
+import { NAVIGATION_PATH } from '~modules/navigation/navigation.consts';
+import { COLLECTIONS_OR_BUNDLES_PATH } from './collection.const';
+import { CONTENT_PATH } from './content.const';
+import { USER_PATH } from './user.const';
 
 export const ADMIN_PATH = Object.freeze({
 	...USER_PATH,
