--- demo/src/react-admin/modules/user/views/index.ts
@@ -1,1 +1,0 @@
-export { default as UserOverview } from './UserOverview';
